requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
langchain-huggingface
huggingface_hub
transformers
accelerate
bitsandbytes
langchain
fastapi
uvicorn
sse_starlette
bs4
pypdf
faiss-cpu
faiss-gpu
